Bug 1622530 - Part 5: Convert TryNoteKind into enum class. r=jorendorff
Differential Revision: https://phabricator.services.mozilla.com/D66906

UltraBlame original commit: 4e13790492c627a8d61d9ae19497446d779846ae

+++ js/src/frontend/BytecodeSection.cpp
@@ -140,7 +140,8 @@ bool CGTryNoteList::append(TryNoteKind kind, uint32_t stackDepth,
   
   
 
-  TryNote note(kind, stackDepth, start.toUint32(), (end - start).toUint32());
+  TryNote note(uint32_t(kind), stackDepth, start.toUint32(),
+               (end - start).toUint32());
 
   return list.append(note);
 }
